requirements.txt and exceptions.py (#55)
* Fixes #34
* Fixes #51

* added requirements.txt, modified README.md, and added exceptions.py

* fixed flake8 in exceptions.py. modified addusers, message, and set-quotas to import from exceptions.py

* changed the name exceptions.py to moc_exceptions.py

* fixed few flake8 errors. Also in message.py, changed 'import moc_exceptions' to 'import .moc_exceptions'
