+++ set-quotas.py
@@ -33,13 +33,7 @@ from moc_utils import get_absolute_path, select_rows
 from quotas import QuotaManager
 from message import TemplateMessage
 import spreadsheet
-
-
-class ProjectNotFoundError(Exception):
-    """The specified project does not exist"""
-    def __init__(self, project_name):
-        message = 'Cannot find project: {}'.format(project_name)
-        super(ProjectNotFoundError, self).__init__(message)
+from moc_exceptions import ProjectNotFoundError
 
 
 def parse_rows(rows, select_project=None):
