+++ tests/test_message.py
@@ -10,7 +10,8 @@ TEST_DIR = os.path.dirname(os.path.abspath(__file__))
 PROJECT_DIR = os.path.abspath(os.path.join(TEST_DIR, os.pardir))
 sys.path.insert(0, PROJECT_DIR)
 
-from message import Message, TemplateMessage, BadEmailRecipient
+from message import Message, TemplateMessage
+from moc_exceptions import BadEmailRecipient
 
 
 def test_template():
